Use apollo hooks for profile tab
